moved vram functions to separate file
